pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "python-dotenv>=1.2.2",
    "snowflake-connector-python>=4.3.0",
    "uvicorn>=0.41.0",
]
